refactor: recommend 가독성 처리

+++ src/main/java/gladiator/philosopher/post/dto/PostResponseDto.java
@@ -1,7 +1,6 @@
 package gladiator.philosopher.post.dto;
 
 import gladiator.philosopher.post.entity.Post;
-import gladiator.philosopher.post.entity.PostImage;
 import gladiator.philosopher.post.entity.PostOpinion;
 import java.time.LocalDateTime;
 import java.util.List;
@@ -19,7 +18,7 @@ public class PostResponseDto {
 
   private final List<String> images;
 
-  private final int recommendCount;
+  private final long recommendCount;
 
   private final List<String> opinions;
 
@@ -29,7 +28,7 @@ public class PostResponseDto {
 
   private final LocalDateTime modifiedDate;
 
-  public PostResponseDto(Post post, int recommendCount, List<String> postImage) {
+  public PostResponseDto(Post post, long recommendCount, List<String> postImage) {
     this.postId = post.getId();
     this.nickname = post.getAccount().getNickname();
     this.title = post.getTitle();
